Adding new file for pi annotated with 4 boxes
Last time I uploaded a file with the same name as the old one, and while the file looks correct in the repository, the published website still shows the old picture. I'm uploading the same file but with a new name so I can change the reference in the lab page.